pyproject.toml
[project]
name = "alexa-skill"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ask-sdk-core>=1.19.0",
    "ask-sdk-webservice-support>=1.3.3",
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
